Add support for setting method call timeout
For some D-Bus services the default D-Bus timeout (25s) may not be
enough; conversely for some applications the default timeout may be
much longer than they'd like to wait.

Add support for setting the timeout to use for method invocations
globally. At least for PHP < 8 this is probably the best we can do as
there is no way to pass additional timeout information to individual
method calls: we cannot reliably tell whether a positional
parameter (to __call()) is the optional timeout or a parameter for the
D-Bus method to invoke.

For PHP 8+ it may be possible to use named arguments to pass a
per-invocation timeout parameter but that's left for the future.